+++ R/tongfen_ca_estimate.R
@@ -67,7 +67,8 @@ tongfen_estimate_ca_census <- function(geometry, meta, level,
 
   # So maybe a function like get_tongfen_correspondence_from_seed
   census_data <- get_tongfen_ca_census(regions = regions, meta = meta,
-                                       level = level, na.rm = na.rm, quiet = quiet) %>%
+                                       level = level, base_geo = datasets,
+                                       na.rm = na.rm, quiet = quiet) %>%
     sf::st_transform(st_crs(geometry))
 
   if (!is.null(downsample_level)){